+++ apps/src/blockly/addons/cdoBlockSvg.js
