+++ apps/src/blockly/addons/cdoMetricsManager.js
