+++ apps/src/blockly/addons/cdoTrashcan.js
