+++ apps/src/blockly/addons/cdoWorkspaceSvg.js
