+++ apps/src/sites/studio/pages/googleblockly.js
@@ -3,6 +3,6 @@ import locale from 'blockly/msg/en';
 import 'blockly/blocks';
 import 'blockly/javascript';
 GoogleBlockly.setLocale(locale);
-import initializeGoogleBlocklyWrapper from './googleBlocklyWrapper';
+import initializeGoogleBlocklyWrapper from '@cdo/apps/blockly/googleBlocklyWrapper';
 
 window.Blockly = initializeGoogleBlocklyWrapper(GoogleBlockly);
