+++ apps/test/unit/blockly/addons/cdoFieldImageDropdownTest.js
@@ -1,5 +1,5 @@
-import {__TestInterface} from '@cdo/apps/blocklyAddons/cdoFieldImageDropdown';
-import {expect} from '../../util/reconfiguredChai';
+import {__TestInterface} from '@cdo/apps/blockly/addons/cdoFieldImageDropdown';
+import {expect} from '../../../util/reconfiguredChai';
 
 describe('CdoFieldImageDropdown', () => {
   it('restructures menu items to make images', () => {
